index exists check

+++ src/services/elasticsearch/elasticsearch.ts
@@ -1,25 +1,36 @@
 import { Client } from "@elastic/elasticsearch";
 
+
 let client = new Client({
   node: process.env.ELASTICSEARCH_NODE
     ? process.env.ELASTICSEARCH_NODE
-    : "http://localhost:9200"
+    : "http://127.0.0.1:9200"
 });
 
 /**
  * Updates elasticsearch with new records
  */
 const updateIndex = async (data: Array<unknown>) => {
-  // First delete all entries
-  await client.deleteByQuery({
-    index: "stock-agg",
-    body: {
-      query: {
-        match_all: {}
-      }
-    }
-  });
+  try {
 
+    // First delete all entries if stock-agg index exists
+
+    const {body} = await client.indices.exists({index: "stock-agg"})
+
+    if (body) {
+      await client.deleteByQuery({
+        index: "stock-agg",
+        body: {
+          query: {
+            match_all: {}
+          }
+        }
+      });
+      console.log("*** Index cleared ***");
+    }
+  } catch (e) {
+    throw Error(`Failed to clear ElasticSearch index: ${e}`);
+  }
   // Now update
   const body = data.flatMap(stock => [
     { index: { _index: "stock-agg" } },
